feat(api): add search module

+++ api-server/src/modules/review/review.service.ts
@@ -3,6 +3,7 @@ import { Review } from '../../entities/review.entity';
 import { AppService } from '../app/app.service';
 import {
   BadRequestException,
+  ForbiddenException,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
@@ -10,6 +11,8 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateReviewDto } from './dto/create-review.dto';
 import { UpdateReviewDto } from './dto/update-review.dto';
+import { JwtUser } from 'src/entities/user.entity';
+import { UserService } from '../user/user.service';
 
 @Injectable()
 export class ReviewService {
@@ -18,20 +21,20 @@ export class ReviewService {
     private readonly reviewRepository: Repository<Review>,
     private readonly keywordService: KeywordService,
     private readonly appService: AppService,
+    private readonly userService: UserService,
   ) {}
 
-  async create({
-    hole,
-    content,
-    keywords,
-    appName,
-    appIconUrl,
-  }: CreateReviewDto) {
+  async create(
+    { hole, content, keywords, appName, appIconUrl }: CreateReviewDto,
+    { id: userId }: JwtUser,
+  ) {
+    const user = await this.userService.findOneById(userId);
     const app = await this.appService.upsert(appName, appIconUrl);
     const review = this.reviewRepository.create({
       hole,
       content,
       app,
+      user,
     });
 
     review.keywords = keywords
@@ -65,10 +68,17 @@ export class ReviewService {
     return review;
   }
 
-  async update(id: number, { hole, content, keywords }: UpdateReviewDto) {
+  async update(
+    id: number,
+    { hole, content, keywords }: UpdateReviewDto,
+    { id: userId }: JwtUser,
+  ) {
     const review = await this.reviewRepository.findOne(id);
     if (!review) {
       throw new NotFoundException();
+    }
+    if (review.user.id !== userId) {
+      throw new ForbiddenException(`Cannot update other user's review`);
     }
 
     return await this.reviewRepository.save({
@@ -85,10 +95,13 @@ export class ReviewService {
     });
   }
 
-  async remove(id: number): Promise<void> {
+  async remove(id: number, { id: userId }: JwtUser): Promise<void> {
     const review = await this.reviewRepository.findOne(id);
     if (!review) {
       throw new NotFoundException();
+    }
+    if (review.user.id !== userId) {
+      throw new ForbiddenException(`Cannot delete other user's review`);
     }
 
     await this.reviewRepository.remove(review);
